universal-idp/requirements.txt
boto3>=1.34.0
PyMuPDF>=1.23.0
pdf2image>=1.16.3
Pillow>=10.0.0
PyPDF2>=3.0.0
pypdfium2>=4.0.0
awscli>=1.29.0
flask>=3.1.0
flask-cors>=4.0.0
pymongo>=4.6.0
sentence-transformers>=2.2.0
numpy>=1.24.0
elasticsearch>=8.0.0